+++ Xen Tuner/debug xen tuner.qml
@@ -31,10 +31,22 @@ import Qt.labs.settings 1.0
 import FileIO 3.0
 
 MuseScore {
-      version: "0.3.1"
-      description: "Create fingerings to display edo/neji-steps of notes relative to the reference note.\n\n" +
-        "Applies to selection, or entire score if nothing is selected."
-      menuPath: "Plugins.Xen Tuner.Display Steps"
+      version: "0.4.0"
+      description: "Debug Tune function.\n\n" +
+        "The docking Xen Tuner plugin is hard to debug as the shortcuts break everytime you re-run " +
+        "the plugin in the plugin creator. Use this instead to test functions without having to " +
+        "restart MuseScore."
+      menuPath: "Plugins.Xen Tuner.Debug Xen Tuner"
+      
+      id: pluginId
+
+      Component.onCompleted : {
+        if (mscoreMajorVersion >= 4) {
+          pluginId.title = qsTr("Xen Tuner");
+          // pluginId.thumbnailName = "some_thumbnail.png";
+          pluginId.categoryCode = "composing-arranging-tools";
+        }
+      }
 
       FileIO {
         id: fileIO
@@ -49,9 +61,15 @@ MuseScore {
       }
 
       onRun: {
-        console.log('Xen Tuner - Display Steps');
-        Fns.init(Accidental, NoteType, SymId, Element, Ms, fileIO, Qt.resolvedUrl("."), curScore);
+        console.log('Xen Tuner - Debug');
+        // When you want to find which import has a syntax error, uncomment this line
+        // console.log(JSON.stringify(Fns));
+        var isMS4 = mscoreMajorVersion >= 4;
+        Fns.init(Accidental, NoteType, SymId, Element,
+          fileIO, Qt.resolvedUrl("../"), curScore, isMS4);
 
-        Fns.operationTune(2);
+        // Debug code here.
+        Fns.operationTune(); // test tune
+        // Fns.operationTranspose(-1, 0); // test diatonic transpose
       }
 }