+++ Xen Tuner/xen tuner.qml
@@ -31,17 +31,36 @@ import Qt.labs.settings 1.0
 import FileIO 3.0
 
 MuseScore {
-      version: "0.3.1"
-      pluginType: "dock"
-      dockArea: "left"
+      version: "0.4.0"
+      pluginType: "dialog" // changed from "dock" to "dialog" to support both MS3 and 4
+      // dockArea: "left"
       description: "Starts the XenTuner plugin.\n\n" +
         "This will open a small docked panel to the side.\n\nIMPORTANT: Do not close the window.\n"+
         "Make sure you only have 1 instance of this plugin running at a time."
       menuPath: "Plugins.Xen Tuner.Start Xen Tuner"
       implicitHeight: 80
-      implicitWidth: 200
+      implicitWidth: 300
       id: pluginId
       readonly property var window: Window.window
+      
+      onRun: {
+        console.log('Started Xen Tuner');
+        // When you want to find which import has a syntax error, uncomment this line
+        // console.log(JSON.stringify(Fns));
+
+        var isMS4 = mscoreMajorVersion >= 4;
+        Fns.init(Accidental, NoteType, SymId, Element,
+          fileIO, Qt.resolvedUrl("../"), curScore, isMS4);
+        console.log('present working dir: ' + Qt.resolvedUrl("../"));
+      }
+
+      Component.onCompleted : {
+        if (mscoreMajorVersion >= 4) {
+          pluginId.title = qsTr("Xen Tuner");
+          // pluginId.thumbnailName = "some_thumbnail.png";
+          pluginId.categoryCode = "composing-arranging-tools";
+        }
+      }
 
       FileIO {
         id: fileIO
@@ -71,7 +90,8 @@ MuseScore {
             font.pointSize: 8
             onClicked: {
                 infoText.text = "Quitting...";
-                Qt.quit();
+                handleClose();
+                pluginId.parent.Window.window.close();
             }
         }
       }
@@ -79,6 +99,7 @@ MuseScore {
       Shortcut {
         sequence: "Alt+R"
         context: Qt.ApplicationShortcut
+        id: tuneShortcut
         onActivated: {
             infoText.text = "Tuning score/selection...";
             Fns.operationTune();
@@ -108,6 +129,7 @@ MuseScore {
       Shortcut {
         sequence: "Alt+Up"
         context: Qt.ApplicationShortcut
+        id: up1Shortcut
         onActivated: {
             infoText.text = "Moving note(s) up aux 1";
             Fns.operationTranspose(1, 1);
@@ -117,6 +139,7 @@ MuseScore {
       Shortcut {
         sequence: "Ctrl+Alt+Up"
         context: Qt.ApplicationShortcut
+        id: up2Shortcut
         onActivated: {
             infoText.text = "Moving note(s) up aux 2";
             Fns.operationTranspose(1, 2);
@@ -126,6 +149,7 @@ MuseScore {
       Shortcut {
         sequence: "Alt+Shift+Up"
         context: Qt.ApplicationShortcut
+        id: up3Shortcut
         onActivated: {
             infoText.text = "Moving note(s) up aux 3";
             Fns.operationTranspose(1, 3);
@@ -135,6 +159,7 @@ MuseScore {
       Shortcut {
         sequence: "Ctrl+Alt+Shift+Up"
         context: Qt.ApplicationShortcut
+        id: up4Shortcut
         onActivated: {
             infoText.text = "Moving note(s) up aux 4";
             Fns.operationTranspose(1, 4);
@@ -145,6 +170,7 @@ MuseScore {
         sequence: "End"
         enabled: false // set to true to enable
         context: Qt.ApplicationShortcut
+        id: up5Shortcut
         onActivated: {
             infoText.text = "Moving note(s) up aux 5";
             Fns.operationTranspose(1, 5);
@@ -155,6 +181,7 @@ MuseScore {
         sequence: "End"
         enabled: false
         context: Qt.ApplicationShortcut
+        id: up6Shortcut
         onActivated: {
             infoText.text = "Moving note(s) up aux 6";
             Fns.operationTranspose(1, 6);
@@ -174,6 +201,7 @@ MuseScore {
       Shortcut {
         sequence: "Alt+Down"
         context: Qt.ApplicationShortcut
+        id: down1Shortcut
         onActivated: {
             infoText.text = "Moving note(s) down aux 1";
             Fns.operationTranspose(-1, 1);
@@ -183,6 +211,7 @@ MuseScore {
       Shortcut {
         sequence: "Ctrl+Alt+Down"
         context: Qt.ApplicationShortcut
+        id: down2Shortcut
         onActivated: {
             infoText.text = "Moving note(s) down aux 2";
             Fns.operationTranspose(-1, 2);
@@ -192,6 +221,7 @@ MuseScore {
       Shortcut {
         sequence: "Alt+Shift+Down"
         context: Qt.ApplicationShortcut
+        id: down3Shortcut
         onActivated: {
             infoText.text = "Moving note(s) down aux 3";
             Fns.operationTranspose(-1, 3);
@@ -201,6 +231,7 @@ MuseScore {
       Shortcut {
         sequence: "Ctrl+Alt+Shift+Down"
         context: Qt.ApplicationShortcut
+        id: down4Shortcut
         onActivated: {
             infoText.text = "Moving note(s) down aux 4";
             Fns.operationTranspose(-1, 4);
@@ -211,6 +242,7 @@ MuseScore {
         sequence: "Home"
         enabled: false
         context: Qt.ApplicationShortcut
+        id: down5Shortcut
         onActivated: {
             infoText.text = "Moving note(s) down aux 5";
             Fns.operationTranspose(-1, 5);
@@ -221,19 +253,12 @@ MuseScore {
         sequence: "Home"
         enabled: false
         context: Qt.ApplicationShortcut
+        id: down6Shortcut
         onActivated: {
             infoText.text = "Moving note(s) down aux 5";
             Fns.operationTranspose(-1, 5);
             afterOperation();
         }
-      }
-
-      onRun: {
-        console.log('Started Xen Tuner');
-        // When you want to find which import has a syntax error, uncomment this line
-        // console.log(JSON.stringify(Fns));
-        Fns.init(Accidental, NoteType, SymId, Element, Ms, fileIO, Qt.resolvedUrl("."), curScore);
-        console.log('present working dir: ' + Qt.resolvedUrl("."));
       }
 
       onScoreStateChanged: {
@@ -263,5 +288,19 @@ MuseScore {
       function afterOperation() {
         // Don't do this, it will steal focus from the score
         // pluginId.window.requestActivate();
+      }
+
+      function handleClose() {
+        console.log('Quitting');
+        var shortcuts = [
+          tuneShortcut, enharmonicShortcut, upShortcut, up1Shortcut,
+          up2Shortcut, up3Shortcut, up4Shortcut, up5Shortcut, up6Shortcut,
+          downShortcut, down1Shortcut, down2Shortcut, down3Shortcut,
+          down4Shortcut, down5Shortcut, down6Shortcut
+        ];
+        for (var i = 0; i < shortcuts.length; i++) {
+          shortcuts[i].context = Qt.WindowShortcut; // make the shortcut disappear with the window.
+          console.log('disable shortcut: ' + shortcuts[i].sequence);
+        }
       }
 }