+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## 0.4.0
+
+- Attempt MuseScore 4 support.
+  - Changed 'dock' to 'dialog' type to support MuseScore 4.
+  - Remove dependency on old `PluginAPI::Ms`, using `Element.name` instead.
+  - Use `AccidentalType` when on MS4 instead of `Accidental`.
+- Fix shortcuts breaking when plugin closed and reopened in plugin creator/MS 4.
+
+## 0.3.2
+
+- Huge performance improvements for tuning & up/down/enharmonic cycle for large scores
+  - Use `Cursor.rewindToTick()` introduced in MS3.5 instead of old method which was very slow.
+  - Improve `getBarBoundaries()` from O(n) to O(log n).
+  - Reduce unnecessary repeated calls to `getBarBoundaries()`, and `removeUnnecessaryAccidentals`
+- Added reference tuning change to change reference nominal without changing the tuning (e.g. `D4:`).
+
 ## 0.3.1
 
 - Added Display Cents and Display Steps plugins to create fingerings to display cent offsets and edo/neji steps of notes in the score.